moved GetFilteredRoutines to separate component
