+++ src/components/MyRoutines.jsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useState } from "react";
 import CreatingRoutines from "./CreatingRoutines";
+import GetFilteredRoutines from "./GetFilteredRoutines";
 import useAuth from "../hooks/useAuth";
 import { getPublicRoutinesByUser } from "../api/routines";
 /**
@@ -15,64 +16,6 @@ import { getPublicRoutinesByUser } from "../api/routines";
  *                   on the routine
  *              [] - be able to remove any activity from the routine
  */
-
-// get Routines again and filter by comparing user.username to creator name
-// to only show the user's routines
-const GetFilteredRoutines = () => {
-  const [userRoutines, setUserRoutines] = useState([]);
-  const { user } = useAuth();
-  //console.log("user:", user.username);
-  useEffect(
-    (username) => {
-      const getAllPublicRoutinesByUser = async () => {
-        const response = await getPublicRoutinesByUser(user.username);
-        setUserRoutines(response);
-      };
-      getAllPublicRoutinesByUser();
-    },
-    [user.username]
-  );
-  return (
-    <div className="routines">
-      {userRoutines && userRoutines.length
-        ? userRoutines.map((userRoutine, id) => {
-            return (
-              <div className="routineList" key={`routine${id}`}>
-                <ul>
-                  <li>
-                    <h2>creator name: {userRoutine.creatorName}</h2>
-                  </li>
-                  <li>
-                    <p>Name: {userRoutine.name}</p>
-                  </li>
-                  <li>
-                    <p>Goal: {userRoutine.goal}</p>
-                  </li>
-                  <li>
-                    <p>Activities: </p>
-                    <ul>
-                      {userRoutine.activities.length
-                        ? userRoutine.activities.map((activity, i) => {
-                            return (
-                              <div key={`activity${i}`}>
-                                <li>Activity Name: {activity.name}</li>
-                                <li>Count:{activity.count}</li>
-                                <li>Duration:{activity.duration}</li>
-                              </div>
-                            );
-                          })
-                        : null}
-                    </ul>
-                  </li>
-                </ul>
-              </div>
-            );
-          })
-        : null}
-      )
-    </div>
-  );
-};
 
 const MyRoutines = ({ post, setPost }) => {
   return (
